refactor: changed type of 'avg_headers_count' and 'avg_query_params_count' from 'Integer' to 'Double'

+++ src/main/java/axi/practice/data_generation_reports/dto/filter/CreateRequestFilterRequestDto.java
@@ -16,6 +16,6 @@ public class CreateRequestFilterRequestDto {
     private LocalDateTime toDate;
     private String host;
     private String path;
-    private Integer avgHeaders;
-    private Integer avgQueryParams;
+    private Double avgHeaders;
+    private Double avgQueryParams;
 }
