+++ src/main/java/axi/practice/data_generation_reports/dto/filter/RequestFilterDto.java
@@ -17,6 +17,6 @@ public class RequestFilterDto {
     private LocalDateTime toDate;
     private String host;
     private String path;
-    private Integer avgHeaders;
-    private Integer avgQueryParams;
+    private Double avgHeaders;
+    private Double avgQueryParams;
 }
